Revert "x86_64: port kernel arch, IPC ABI, drivers, servers, and userland to amd64"
This reverts commit 0457e10bcd34a4f578da59c5f3c95cb27ce7912c.

+++ minix/commands/part/part.c
@@ -791,7 +791,7 @@ void print(object_t *op)
 		sprintf(op->value, "%s%d",
 			curdev->parttype >= PRIMARY ? "p" :
 				curdev->parttype == SUBPART ? "s" : "",
-			(int)(curdev->parttype == SUBPART ||
+			(curdev->parttype == SUBPART ||
 				curdev->parttype == FLOPPY ? pe - table
 					: sort_index[pe - table]) - 1);
 		break;
