+++ minix/fs/mfs/super.c
@@ -322,9 +322,9 @@ int read_super(struct super_block *sp)
   	return(EINVAL);
   }
 
-  /* Limit s_max_size to INT32_MAX */
-  if ((unsigned long)sp->s_max_size > INT32_MAX)
-	sp->s_max_size = INT32_MAX;
+  /* Limit s_max_size to LONG_MAX */
+  if ((unsigned long)sp->s_max_size > LONG_MAX) 
+	sp->s_max_size = LONG_MAX;
 
   sp->s_isearch = 0;		/* inode searches initially start at 0 */
   sp->s_zsearch = 0;		/* zone searches initially start at 0 */
